+++ NEWS.md
@@ -1,4 +1,4 @@
-# semptools 0.2.11.1
+# semptools 0.2.11.3
 
 ## New Features
 
@@ -18,6 +18,33 @@
   the plot to fit the rectangle bounded
   by -1 and 1 vertically and
   horizontally. (0.2.11.1)
+
+## Improvement
+
+- Updated `mark_sig()` to accept a
+  data frame with the *p*-values. Users
+  can supply *p*-values computed by
+  other functions. (0.2.11.3)
+
+- Updated `mark_se()` to accept a
+  data frame with the standard errors.
+  Users can supply standard errors
+  computed by other functions. (0.2.11.3)
+
+- Updated `mark_sig()` to use the *p*-values
+  from `lavaan::standardizedSolution()`.
+  Users need to explicitly request
+  standardized solution *p*-values by
+  setting the argument `std_type`.
+  (0.2.11.3)
+
+- Updated `mark_se()` to use the
+  standard errors
+  from `lavaan::standardizedSolution()`.
+  Users need to explicitly request
+  standardized solution standard errors
+  by setting the argument `std_type`.
+  (0.2.11.3)
 
 ## Miscellaneous